refactoring

+++ mcl_2d/include/mcl_2d/msg_converter.hpp
@@ -22,8 +22,8 @@ class MsgConverter {
   geometry_msgs::msg::PoseArray createParticleCloud(std::vector<Mcl2d::Particle>& particles);
   sensor_msgs::msg::PointCloud2 createTransformedPC2(Eigen::Matrix4Xf& eigenLaser, Eigen::Matrix4f& transMatrix);
 
-  vector<LaserPoint> scan_to_vector(const sensor_msgs::msg::LaserScan::SharedPtr msg, const Vector3d& laser);
+  vector<LaserPoint> scan_to_vector(const sensor_msgs::msg::LaserScan::SharedPtr msg, const Vector3f& laser);
   sensor_msgs::msg::PointCloud2 vector_to_PC2(vector<LaserPoint>& points);
-  geometry_msgs::msg::TransformStamped broadcastWorldToBaseLink(const Vector3d& pose);
-  geometry_msgs::msg::TransformStamped broadcastBaseLinkToLidarFrame(const Vector3d& pose);
+  geometry_msgs::msg::TransformStamped broadcastWorldToBaseLink(const Vector3f& pose);
+  geometry_msgs::msg::TransformStamped broadcastBaseLinkToLidarFrame(const Vector3f& pose);
 };
